backend/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "Template project for a fullstack web app with LLM support"
authors = ["Johan Berggren <jberggren@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
Authlib = "^1.3.0"
itsdangerous = "^2.1.2"
SQLAlchemy = "^2.0.27"
requests = "^2.31.0"
google-cloud-aiplatform = "^1.43.0"
httpx = "^0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.group.dev.dependencies]
pylint = "^3.1.0"
pytest = "^8.0.2"
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
